% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bs2pb.R
\name{boost2protobuf}
\alias{boost2protobuf}
\title{convert the GatingSet from boost to protobuf format}
\usage{
boost2protobuf(from, to, ...)
}
\arguments{
\item{from}{the old archive path}

\item{to}{the new archive path}

\item{...}{other arguments passed to \link{save_gs}}
}
\description{
convert the GatingSet from boost to protobuf format
}
\examples{
\dontrun{
#set cdf to 'link' to avoid the redundant copy of flow data
boost2protobuf(old_gs_path, new_gs_path, cdf = "link")
}
}

