% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_gs.R
\name{load_gslist}
\alias{load_gs}
\alias{load_gslist}
\title{load a GatingSet/GatingSetList from disk.}
\usage{
load_gslist(path)

load_gs(path)
}
\arguments{
\item{path}{A character scalar giving the path to load the GatingSet from.}
}
\value{
\code{load_gs} returns a GatingSet object
\code{load_gslist}_bs returns a GatingSetList object
}
\description{
It is used to load the legacy GatingSet saved in boost serialization format.
}
\examples{
\dontrun{
	gs <- boost2protobuf:::load_gs(path="tempFolder")
 # save it in pb format
	boost2protobuf:::save_gs(gs,path="newFolder", cdf = "link)
}
}

