% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_gs.R
\name{save_gslist}
\alias{load_gs}
\alias{load_gslist}
\alias{save_gs}
\alias{save_gslist}
\title{save a boost GatingSet/GatingSetList as protocol buffer format}
\usage{
save_gslist(gslist, path, ...)

save_gs(G, path, overwrite = FALSE, cdf = c("copy", "move", "skip",
  "symlink", "link"), ...)
}
\arguments{
\item{gslist}{A \code{GatingSetList}}

\item{path}{A character scalar giving the path to save/load the GatingSet to/from.}

\item{...}{other arguments: not used.}

\item{G}{A \code{GatingSet}}

\item{overwrite}{A logical scalar specifying whether to overwrite the existing folder.}

\item{cdf}{a character scalar. The valid options are :"copy","move","skip","symlink","link" specifying what to do with the cdf data file.
Sometime it is more efficient to move or create a link of the existing cdf file to the archived folder.
It is useful to "skip" archiving cdf file if raw data has not been changed.}
}
\value{
\code{load_gs} returns a GatingSet object
\code{load_gslist} returns a GatingSetList object
}
\description{
Save a GatingSet/GatingSetList to the disk using protocol buffer library.
}
\examples{
\dontrun{
	gs<- boost2protobuf:::load_gs(bsFolder)
 boost2protobuf:::save_gs(pbFolder)

}
}
\seealso{
\code{\link{GatingSet-class}},\code{\link{GatingSetList-class}}
}

